Added importing and exporting of rigged and animated COLLADA models (non-complex models). Static models eg. level models can also be exported to COLLADA. Added animation viewer supporting ~90% of SM64DS animations. Added support for exporting vertex colours in OBJ using Blender-specific "Extended OBJ" plugin. Improved how collision map appears in KCL editor. Added selecting faces by clicking in KCL editor. Fixed bug that produced incorrect faces for very large faces in collision maps. Fixed bug in level editor when adding and removing path nodes. OpenTK version updated to 1.1.1536.5788. Added 3 new additional patches: 1) "All Objects Always Visible from a Distance" by Skelux, 2) "Increase Star Object Draw Distance (USv1)" by skawo, 3) "All TTC Objects Operate at a Fixed Speed Regardless of When You Enter the Level" by Skelux.

+++ Importers/KCL_Importer.cs
@@ -679,9 +679,9 @@ namespace SM64DSe
             this.scale = scale;
         }
 
-        public int valToWrite()
+        public long valToWrite()
         {
-            return (int)(theValue / scale);
+            return (long)(theValue / scale);
         }
 
         public int sizeOf()
